Former-commit-id: 6277d9f2254376f8d345642f61300b2c0419ecae
Former-commit-id: e127f9515c73006386241f90149ba8268baf2e44

+++ src/RiakClient/Models/Search/RiakSearchRequest.cs
@@ -53,9 +53,9 @@ namespace RiakClient.Models.Search
     {
         public RiakFluentSearch Query { get; set; }
 
-        public uint Rows { get; set; }
+        public long Rows { get; set; }
 
-        public uint Start { get; set; }
+        public long Start { get; set; }
 
         public RiakFluentSearch Filter { get; set; }
 
@@ -80,8 +80,8 @@ namespace RiakClient.Models.Search
             {
                 index = Query.Index.ToRiakString(),
                 q = Query.ToString().ToRiakString(),
-                rows = Rows,
-                start = Start,
+                rows = (uint)Rows,
+                start = (uint)Start,
                 sort = Sort.ToRiakString(),
                 filter = Filter == null ? null : Filter.ToString().ToRiakString(),
                 presort = PreSort.HasValue ? PreSort.Value.ToString().ToLower().ToRiakString() : null,
